LPS-109189 Move IndexUpdaterUtil to a more specific package, com.liferay.portal.db.index
